Implemented per-page or site-wide background images.
If the user sets "background: /path/to/image.jpg" or an http(s) url in their _config.yml, or if they set it in a post/page's yaml front matter, the displayed page will use that image as the background. If their background is transparent, the default background color will also be visible. Setting a background in a page/post will take precedence over the one set in _config.yml.